remove more nocommits; small renaming/refactoring; optimize termsenum seeking for flex API on pre-flex index; new bulk read api
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/branches/flex_1458@921608 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/lucene/index/codecs/preflex/SegmentTermEnum.java
@@ -129,7 +129,7 @@ public final class SegmentTermEnum extends TermEnum implements Cloneable {
     return clone;
   }
 
-  final void seek(long pointer, int p, Term t, TermInfo ti)
+  final void seek(long pointer, long p, Term t, TermInfo ti)
           throws IOException {
     input.seek(pointer);
     position = p;
